Refactor TransactionServiceImplementation with getUser and findAccount methods
This commit adds the getUser and findAccount methods to the TransactionServiceImplementation class. The getUser method decodes the token and sends an HTTP GET request to the user service to retrieve the user details. The UserDetailsResponse record was created to hold the response from the user service.

The findAccount method searches the user's account based on the account number in the request. The UserAccounts record was created for the user account data.

The transact method is split further by creating the createTransaction method, which creates a new transaction object and updates the account balance based on the transaction type.

The updated TransactionResponse object contains the transaction details and the updated account balance.

The update was tested, it works as it should. No part of the codebase is broken.