Add Kafka configuration to Transaction Service
This commit adds Kafka configuration to the Transaction Service. Specifically, the following changes have been made:
- Added spring-kafka dependency to pom.xml
- Created KafkaConfig class with configuration for producing and consuming messages using Kafka
- Added Kafka properties to application.properties file to configure the Kafka connection and serialization

These changes enable the Transaction Service to communicate with other services using Kafka, and ensure that messages are properly serialized and deserialized. The KafkaConfig class contains the necessary beans for producing and consuming messages, and the application.properties file contains the necessary properties to configure the Kafka connection and serialization.

Note that a new topic, "transaction-creation", has also been added to the kafka.topic section of the properties file.

This commit is part of ongoing work to enable inter-service communication in the Transaction Service.

This commit also refactors the apiConfig folder to config.