Add Kafka messaging functionality and TransactionMessage class to transact method
This commit adds the necessary code to send transaction messages to other services via Kafka. Specifically, adds the code to the transact method to create a new TransactionMessage object and send it to the transactionCreationTopic via the kafkaTemplate.

Changes:
- Added code to transact method to send transaction messages via Kafka
- Used kafkaTemplate.send to send the transaction message to the transactionCreationTopic

The transact method was tested and it works as should. Messages are being sent correctly for every transaction. No part of the codebase is broken